fix(move-app): split block-writes integration test to avoid vi.mock collision
